feat: use @solidjs/router and rename Loading to LoadingSpinner
