+++ apps/client/src/App.tsx
@@ -1,24 +1,15 @@
 import { AuthContext } from "@mp/auth/client";
 import { QueryClientProvider } from "@tanstack/solid-query";
-import { ErrorBoundary, lazy, Suspense } from "solid-js";
-import Layout from "./ui/Layout";
+import { Router } from "@solidjs/router";
 import { authClient, queryClient } from "./state/api";
-import { ErrorFallback } from "./ui/ErrorFallback";
-import { Loading } from "./ui/Loading";
-
-const GamePage = lazy(() => import("./pages/game/GamePage"));
+import Layout from "./ui/Layout";
+import { routes } from "./routes";
 
 export default function App() {
   return (
     <AuthContext.Provider value={authClient}>
       <QueryClientProvider client={queryClient}>
-        <ErrorBoundary fallback={ErrorFallback}>
-          <Layout>
-            <Suspense fallback={<Loading />}>
-              <GamePage />
-            </Suspense>
-          </Layout>
-        </ErrorBoundary>
+        <Router root={Layout}>{routes}</Router>
       </QueryClientProvider>
     </AuthContext.Provider>
   );
