+++ apps/client/src/ui/Layout.tsx
@@ -1,15 +1,15 @@
 import type { ParentProps } from "solid-js";
 import { ErrorBoundary, Suspense } from "solid-js";
 import AppBar from "./AppBar";
+import { LoadingSpinner } from "./LoadingSpinner";
 import { ErrorFallback } from "./ErrorFallback";
-import { Loading } from "./Loading";
 
 export default function Layout(props: ParentProps) {
   return (
     <>
       <AppBar />
       <ErrorBoundary fallback={ErrorFallback}>
-        <Suspense fallback={<Loading />}>{props.children}</Suspense>
+        <Suspense fallback={<LoadingSpinner />}>{props.children}</Suspense>
       </ErrorBoundary>
     </>
   );
