+++ apps/client/src/ui/Link.tsx
@@ -1,6 +1,7 @@
-import type { JSX } from "solid-js";
+import type { AnchorProps } from "@solidjs/router";
+import { A } from "@solidjs/router";
 import * as styles from "./Link.css";
 
-export function Link(props: JSX.IntrinsicElements["a"]) {
-  return <a classList={{ [styles.link]: true }} {...props} />;
+export function Link(props: AnchorProps) {
+  return <A classList={{ [styles.link]: true }} {...props} />;
 }
